Make list of enums from TClass thread safe
The list of enums from TClass are not handled immutably. If the list
already is filled then we return that one. If the list is not filled
and we are told not to load it then we return an empty list. Else
we create a new list and fill it, cache that value and then return it.
In this way once the list is handed out to a caller it is not changed.