Made calling TGenCollectionStreamer::Generate thread safe
Although each thread will get its own copy of TGenCollectionStreamer,
that copy is made by calling Generate on a global instance of the class.
Therefore Generate needs to be thread safe. This translates to requiring
TGenCollectionProxy::Initialize be thread safe. This required that
TGenCollectionProxy::fValue be atomic and be the last value to be set
in TGenCollectionProxy::InitializeEx.