Fix ROOT-9173: add per object flag for the statsoverflow
This change is required to prevent non reproducible behaviours in a multithreaded environments.
For example, if several threads change the TH1::fgStatOverflows flag and fill histograms, the behaviour
will be undefined.

A getter and a setter were added to the TH1 class: TH1::GetStatOverflows and TH1::SetStatOverflows.
The act on a persistent flag, fStatOverflows, which is a property of the histogram object.
The flag of a object is set to the value of the global value TH1::fgStatOverflows at construction time.
Every check done on the TH1::fgStatOverflows flag (also on inherited classes, namely TH1K, TH2, TH3, TProfile,
TProfile2/3D) has been replaced with one done on the data member.
The class version has been increased to 8.