Revert "Reland "[cxxmodules] Don't complain about redeclaration of declared annotated enum."""
This reverts commit bc460a284a7307e54a54a036442cdafc7dd56f5d.

This patch is not sufficient to fix the cmssw issue it intends. We have not yet
created the `New` declaration and we cannot access its annotations. We can only
indirectly assume the conditions from the passed in arguments.

The new patch check will work but ignore possibly more cases which it should not.